Revised casting.

+++ src/src/image_texture.cpp
@@ -14,8 +14,8 @@ ImageTexture::ImageTexture(unsigned char* pixels, int w, int h) :
 
 vec3 ImageTexture::value(float u, float v, const vec3& p) const
 {
-	int i{ u*width };
-	int j{ (1.0f - v)*height - 0.001f };	// invert direction of y axis
+	int i{ (int)(u*width) };
+	int j{ (int)((1.0f - v)*height - 0.001f) };	// invert direction of y axis
 	if (i < 0)	i = 0;
 	if (j < 0)j = 0;
 	if (i > width - 1)i = width - 1;
